requirements-local.txt
requests>=2.31.0
trafilatura>=1.12.0
PyYAML>=6.0
